Refactor into modular package with API
