+++ README.md
@@ -1,50 +1,52 @@
 # WiFi Fixer Tool
 
-Esta aplicación de escritorio está pensada para diagnosticar y resolver problemas de conectividad WiFi y de red en cualquier sistema operativo. Incluye pruebas de gateway, detección de conflictos de IP, verificación de DHCP y chequeo de puertos.
+Esta aplicación de escritorio diagnostica y soluciona problemas de conectividad WiFi y de red en Windows, macOS y Linux. Incluye comprobaciones de gateway, detección de conflictos de IP, verificación de DHCP y test de puertos.
+
+## Estructura del proyecto
+
+El código se organiza como un paquete Python llamado `wifi_fix` que contiene:
+
+- **core.py**: lógica de diagnóstico y reparación.
+- **system.py**: abstracción de comandos específicos por plataforma.
+- **logger.py**: generación de reportes en `wifi_fix_report.txt`.
+- **gui_tkinter.py**: interfaz gráfica simple para escritorio.
+- **server.py**: servicio opcional FastAPI para exponer la lógica a otras interfaces (por ejemplo Tauri/Electron).
+
+El script `wifi_fix_tool.py` ejecuta la interfaz Tkinter por defecto.
 
 ## Requisitos para desarrolladores
-- Python 3 y `pip` para generar el ejecutable.
-- Algunos comandos pueden requerir permisos de administrador/superusuario.
+
+- Python 3.11 o superior.
+- Algunas acciones pueden requerir permisos de administrador/superusuario.
 
 ## Uso
 
-Durante el desarrollo puedes ejecutar el script directamente:
+Durante el desarrollo puedes iniciar la GUI directamente:
 
 ```bash
 python wifi_fix_tool.py
 ```
 
-Se abrirá una ventana con botones para:
-- Verificar la conexión a Internet.
-- Reiniciar el adaptador de red.
-- Renovar la IP.
-- Limpiar la caché DNS.
-- Cambiar a DNS públicos (Google).
-- Ejecutar un **Diagnóstico avanzado** para detectar conflictos de IP, problemas de DHCP y bloqueo de puertos.
-- Ejecutar todas las acciones automáticamente con **Arreglar todo**.
-- Abrir el archivo de reporte generado.
+Para ejecutar la API REST (útil para integrarse con una interfaz Tauri o Electron):
 
-Cada acción genera un reporte en un archivo `wifi_fix_report.txt` ubicado en el directorio personal del usuario.
+```bash
+uvicorn wifi_fix.server:app --reload
+```
+
+Cada acción genera un reporte en `wifi_fix_report.txt` dentro de tu directorio personal.
 
 ## Distribución sin dependencias
 
-Para generar un ejecutable que no requiera tener Python instalado se puede utilizar **PyInstaller**. Una vez instalado, ejecuta:
+Se recomienda usar **PyInstaller** para crear un ejecutable autónomo:
 
 ```bash
 pyinstaller --onefile wifi_fix_tool.py
 ```
 
-El archivo resultante en la carpeta `dist` funcionará con doble clic en Windows, macOS o Linux, dependiendo del sistema donde se genere.
-Los usuarios finales solo necesitan descargar este archivo y ejecutarlo; no es necesario instalar Python, Git ni otras dependencias.
+El archivo resultante en `dist` funcionará con doble clic en el sistema donde se genere. Los usuarios finales no necesitan instalar Python ni otras dependencias.
 
-## Empaquetado automático
-
-Se incluye un script `make_executable.py` para facilitar la generación del ejecutable.
-Asegúrate de tener instalada la herramienta **PyInstaller** y ejecuta:
+También puedes utilizar `make_executable.py` para simplificar la generación del binario:
 
 ```bash
 python make_executable.py
 ```
-
-El binario resultante quedará en la carpeta `dist` y podrá ejecutarse con doble clic
-sin necesidad de instalar Python ni otras dependencias.
